Beginning of the Editor UI
Basic Editor settings being displayed.  Data is not hooked up to any functionality yet.  Will start that with creating a true empty scene that only contains a Grid by default.
Grid will be a special type of Render object.  It is not drawn as a quad but a line list instead.
At same time, enable wireframe capability cross render components